+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Histórico de cambios
 >Todos los cambios notables de este proyecto se documentarán en este archivo.
 
+## [v0.0.3] - 2021-02-09
+### Añadidos
+* Nueva librería ***Mailing.php*** para creación de mails.
+
+### Corrección de errores
+* Reparación de método ***cleanDirectory*** en ***FileFunctions.php***.
+
+---
 ## [v0.0.2] - 2021-01-30
 ### Añadidos
 * Nuevas funciones en ***StringFunctions.php***.